backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
googlemaps==4.10.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0